Refactor : 매서드명 변경
